==> experiments/02_add_time_marks_processed.ipynb <==
# Create one dataset for time marks

- get time data organized by each serving configuration
- run {script}.py to go trough all experiment runs

- Input:
  - results_*/runall_timestamps.csv
  - results_*/load_times.csv
- Output
  - results_*/processed/time_marks.csv
 
[NOV] add results_*/processed/time_marks.csv from different models added
- Input
  - results_*/processed/time_marks.csv -> from original_dir
- Output
  - results_*/processed/time_marks.csv -> from new_dir



pwd

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os

SAVE_TABLES = True

#device = 'gpu'
#new_directory = f"D:/GAISSA/energy-repo/last_repo/june_{device}/" # change
#new_directory = f"D:/GAISSA/energy-repo/last_repo/june_{device}_test/" # change
#new_results = f"D:/GAISSA/energy-repo/last_repo/{device}_june_phi2_01/" # change


#original_results_dir = f"D:/GAISSA/energy-repo/repo_sept_last_update/cudaep_nov_results/nov_cudaep_00/" # change
#new_results_dir = f"D:/GAISSA/energy-repo/repo_sept_last_update/cudaep_nov_results/nov_cudaep_03/" # change if new dataset

original_results_dir = f"D:/GAISSA/energy-repo/repo_sept_last_update/cpuep_nov_results/nov_cpuep_00/" # change
new_results_dir = f"D:/GAISSA/energy-repo/repo_sept_last_update/cpuep_nov_results/nov_cpuep_02/" # change if new dataset
os.chdir(original_results_dir)

os.chdir(original_results_dir)
for i in range(1,11):
    results_dir = f'results_{i}/' 
    print(f"++++++++--------------------results_dir: {results_dir}---------------------------++++++++:")

    os.chdir(original_results_dir)
    original_df = pd.read_csv(results_dir + 'processed/time_marks_updated.csv', index_col=None) # change if new dataset, first: time_marks, then: time_marks_updated
    print(original_df.columns)
    original_df.tail()

    try:
        original_df = original_df.drop(['Unnamed: 0'], axis=1)
    except:
        print("Already deleted")
    print(original_df.columns)
    original_df.head()

    os.chdir(new_results_dir)
    new_df = pd.read_csv(results_dir + 'processed/time_marks.csv', ) #same
    print(new_df.columns)
    new_df.tail()

    # Check if the last row in the "model" column has the value "FINISH"
    if original_df.iloc[-1]['model'] == "FINISH":
        # Remove the last column
        original_df = original_df.iloc[:-1]
        print("Last column removed as the last row in 'model' had the value 'FINISH'.")
        
    else:
        print("No changes made: The last row in 'model' did not contain the value 'FINISH'.")
    #original_df

    # Check if the last row in the "model" column has the value "FINISH"
    if new_df.iloc[0]['model'] == "IDLE":
        # Remove the last column
        new_df = new_df.iloc[1:]
        print("IDLE deleted")
        
    else:
        print("No changes made: The first row in 'model' did not contain the value 'IDLE'.")
    #new_df

    print(len(original_df))
    print(len(new_df))
    combined_df = pd.concat([original_df, new_df], )
    print(len(combined_df))
    
    #combined_df

    os.chdir(original_results_dir)

    processed_updated_name = 'processed/time_marks_updated.csv' # same
    if SAVE_TABLES:
        combined_df.to_csv(results_dir+processed_updated_name, index=False)
        print(f"Results saved in {original_results_dir+results_dir+processed_updated_name}")
    else:
        print(f"Results NOT saved in {original_results_dir+results_dir+processed_updated_name}")
